+++ releases/eucalyptus/3/wb/CHANGELOG.md
@@ -9,6 +9,15 @@ release.
 
 ## [Unreleased]
 
+### Added
+
+- Add `django-redis-sentinel-redux` to allow the use of Redis Sentinel for
+  Django cache
+
+### Changed
+
+- Configure most Django cache backends to Redis
+
 ## [eucalyptus.3-wb-1.8.1] - 2020-04-17
 
 ### Changed